requirements.txt
gostcrypto==1.2.5
pillow==11.0.0
